rewrite it all to work our way
